+++ src/components/Code/Code.style.tsx
@@ -1,5 +1,5 @@
 import { styled } from '@storybook/theming';
-import { getTextColor, getBorderColor } from '../../helper/color';
+import { getTextColor, getBorderColor } from '../../helper';
 
 const StyledCode = styled.div`
   margin: 1.5rem 1rem;