+++ src/components/Panel/Panel.tsx
@@ -12,44 +12,41 @@ import StyledPanel from './Panel.style';
 
 const Panel = () => {
   const {
-    theme,
     themes,
-    activeTheme,
+    activeTheme: { theme, name },
     updateActiveTheme,
     config
   } = React.useContext(SettingsContext);
 
   return (
-    <>
-      <StyledPanel>
-        {theme && Object.entries(theme).length > 0 && (
-          <>
-            <div className="panel__content">
-              {themes.length > 1 && (
-                <div className="panel__themes">
-                  <RadioGroup
-                    label="Active Theme"
-                    name="themes"
-                    value={activeTheme}
-                    onChange={val =>
-                      updateActiveTheme(themes.filter(t => t.name === val)[0])
-                    }
-                  >
-                    {themes.map(t => (
-                      <RadioOption key={t.name} label={t.name} value={t.name} />
-                    ))}
-                  </RadioGroup>
-                </div>
-              )}
-              <SettingsItem obj={theme} arr={[]} />
-              {config.showCode && <Code value={theme} />}
+    <StyledPanel>
+      {Object.entries(theme).length > 0 && (
+        <>
+          <div className="panel__content">
+            {themes.length > 1 && (
+              <div className="panel__themes">
+                <RadioGroup
+                  label="Active Theme"
+                  name="themes"
+                  value={name}
+                  onChange={val =>
+                    updateActiveTheme(themes.filter(t => t.name === val)[0])
+                  }
+                >
+                  {themes.map(t => (
+                    <RadioOption key={t.name} label={t.name} value={t.name} />
+                  ))}
+                </RadioGroup>
+              </div>
+            )}
+            <SettingsItem />
+            {config.showCode && <Code value={theme} />}
 
-              <Loading />
-            </div>
-          </>
-        )}
-      </StyledPanel>
-    </>
+            <Loading />
+          </div>
+        </>
+      )}
+    </StyledPanel>
   );
 };